basic heartbeat - still needs interfacing to monitoring task. there still are un-explained effects (main task causing delays, ping response taking 10x normal message send) which indicates interference with/pitfalls of the tokio scheduler, but general ping response times of 5 micro secs are more than acceptable every 30 secs or so. The variance in response times however indicates we are mostly measuring the scheduler and hence those numbers are only good as relative values (and to check if actors did respond within a check cycle)
tasks are now named so that we can see them in the tokio-console. This
requires to spawn them through task::Builder (which is still
experimental but most likely going to stay). Acceptable since not
crucial.

timers and spawns are now returning Result - task spawning through
builders can fail and we shouldn't panic in this case