extended sentinel alarm server to use a list of (boxed) messengers. Added respective sentinel_alarm cli options. Changed AlarmMessenger to use alarm references and used async_trait to make it object safe
added a odin_build/ as a place to put common build.rs functions. For
server assets we need a build script in every crate that provides them
so we have to avoid code duplication.

Started to move some assets over from RACE to test assets. All the JS
modules still need to be modified