added admin modue to odin_common that allows to build apps so that they can log sysadmin messages (e.g. panics). If built with feature "slack_admin" this posts to a slack channel that is configured in slack_admin.ron
the LiveSentinelConnector can now reconnect if the websocket is for some
reason dropped. This is a major rewrite that uses a single task to
read/write the websocket, including sending ping messages if there was
no websock traffic for a configured amount of time (and not
indiscriminantly as before). This uses Tokio's select, and here it is
actually useful to drop the sleep future if the read/write futures make
progress. The not so nice part was how to keep enough information about
the record wave-front in the task so that we can query the missed ones
upon reconnection