added generic WsService
slight but consequential change in SpaService interface - we now pass in
a &mut SpaConnection into SpaService::init_connection so that we don't
run into the danger of overloading/blocking the server mailbox. It's
also more efficient to directly send to the websocket since this all
happens sync from within the AddConnection processing

switched to proper unicode chars in ui.css (it was suddenly failing in
Chrome).