first functional version of interactive odin_share (geospatial display of selected items still missing)
added intrinsic main.js module to SpaComponents to guarantee that all
modules will see window.main

main.share is now always guaranteed to be set in module postInitialize.
Any use of the main.share functions before would result in exceptions if
respective modules don't have an explicit dependency on an share module.
This is somewhat of a design outlier but we do need an abstract
mechanism to share without each JS module having to hardcode the
dependency on a specific implementation