switched tungstenite to native-tls (rustls was not working with wss:// on macOS)
refactored inactive detection in odin_sentinel. The inactive config does
not belong into sentinel.ron - it is depending on respective
applications. For show_sentinels this is now configured in the
odin_sentinel_config.js as the only processing is on the client. The
snapshot was already sending the list of monitored devices (to detect
the ones that are never updated). For sentinel_alarm this requires to
add a init action to the store which sends the current sentinel states
(last time_recorded for each known device). The duration is configured
in sentinel_alarm.ron (still needs testing)