fixed async module loading problem in Safari - any reference to a odin_cesium.js export prior to its definition seems to hang the referencing module. This is esp. critical for the global Cesium viewer object - we either have to make sure we don't await anything before it is initialized or that referencing modules use the exported viewerReadyPromise to obtain the viewer.
changed topo terrain init in odin_cesium.js to sync (awaiting the
promise) - this avoids flicker and we shouldn't force the use of
promises outside of module init. Otherwise we end up with
promise-spaghetti.

Note this isn't completely safe yet as there can be other var references
for which we don't have a promise yet.

There always will be awaits in odin_cesium if we use async terrain

Made UI settings optional - the icon/window have to be requested (e.g.
in odin_cesium.js). This also takes care of the setting icon order
(it shouldn't be the first).

Fixed odin_adsb visibility control