changes to the JS module init sync - everything that adds Cesium DataSources first has to await the odin_cesium.viewerReadyPromise
added a main.postInitPromises list, to which modules can add promises
that need to be resolved before the postInitialize hooks are called.
This is required for modules that spawn promises and then finish their
init

added a known-problems list.

This currently contains the Safari/VPN
CORS problem, for which we provide the odin_cesium_config.js
'fakeTopoTerrain' option as a temporary workaround. Please note this
only helps with terrain. The Cesium WorldTerrain (Bing) runs into the
same problem and cannot be used in Safari when on VPN