fixed terrain clipped billboard symbols for sentinels, n5 and alertca (needs large disableDepthTestDistance)
added bounded 30sec alertca all_cameras retry if initial file download
fails. It seems the file is quite often in flight at which point the
server just rejects the request

added more color and logo options to mkslides, which now supports
setting dark schemes (if images allow it). The logo (if specified)
is added to the right (float) of H2 slide titles as a ::after element
but is NOT scaled so the image has to have the right size